+++ sdk/core/perf/test/inc/azure/perf/test/win_http_client_get_test.hpp
@@ -11,7 +11,7 @@
 
 #include "azure/perf/test/http_client_get_test.hpp"
 
-#include <azure/core/http/winhttp/win_http_client.hpp>
+#include <azure/core/http/win_http_transport.hpp>
 
 namespace Azure { namespace Perf { namespace Test {