+++ sdk/storage/azure-storage-common/inc/azure/storage/common/shared_key_policy.hpp
@@ -6,7 +6,7 @@
 #include <memory>
 #include <string>
 
-#include <azure/core/http/policy.hpp>
+#include <azure/core/http/policies/policy.hpp>
 
 #include "azure/storage/common/storage_credential.hpp"